Move non-mlmd utils to common module
