+++ src/common/Utils.test.ts
@@ -1,0 +1,63 @@
+import {formatDateString, logger, titleCase} from "./Utils";
+
+describe('Utils', () => {
+  describe('log', () => {
+    it('logs to console', () => {
+      // tslint:disable-next-line:no-console
+      const backup = console.log;
+      global.console.log = jest.fn();
+      logger.verbose('something to console');
+      // tslint:disable-next-line:no-console
+      expect(console.log).toBeCalledWith('something to console');
+      global.console.log = backup;
+    });
+
+    it('logs to console error', () => {
+      // tslint:disable-next-line:no-console
+      const backup = console.error;
+      global.console.error = jest.fn();
+      logger.error('something to console error');
+      // tslint:disable-next-line:no-console
+      expect(console.error).toBeCalledWith('something to console error');
+      global.console.error = backup;
+    });
+  });
+
+  describe('formatDateString', () => {
+    it('handles an ISO format date string', () => {
+      const d = new Date(2018, 1, 13, 9, 55);
+      expect(formatDateString(d.toISOString())).toBe(d.toLocaleString());
+    });
+
+    it('handles a locale format date string', () => {
+      const d = new Date(2018, 1, 13, 9, 55);
+      expect(formatDateString(d.toLocaleString())).toBe(d.toLocaleString());
+    });
+
+    it('handles a date', () => {
+      const d = new Date(2018, 1, 13, 9, 55);
+      expect(formatDateString(d)).toBe(d.toLocaleString());
+    });
+
+    it('handles undefined', () => {
+      expect(formatDateString(undefined)).toBe('-');
+    });
+  });
+
+  describe('titleCase', () => {
+    it('Capitalizes the first letter of each word in a sentence', () => {
+      expect(titleCase('this is a sentence')).toBe('This Is A Sentence');
+    });
+
+    it('Capitalizes the first letter of words seperated by non-word characters',
+      () => {
+        expect(titleCase('test-hyphen_underscore.period'))
+          .toBe('Test Hyphen Underscore.period');
+      });
+
+    it('Returns an empty string when given an empty string',
+      () => {
+        expect(titleCase('')).toBe('');
+      });
+  });
+})
