+++ src/common/Utils.ts
@@ -1,0 +1,25 @@
+export const logger = {
+  error: (...args: any[]) => {
+    // tslint:disable-next-line:no-console
+    console.error(...args);
+  },
+  verbose: (...args: any[]) => {
+    // tslint:disable-next-line:no-console
+    console.log(...args);
+  },
+};
+
+export function formatDateString(date: Date | string | undefined): string {
+  if (typeof date === 'string') {
+    return new Date(date).toLocaleString();
+  } else {
+    return date ? date.toLocaleString() : '-';
+  }
+}
+
+/** Title cases a string by capitalizing the first letter of each word. */
+export function titleCase(str: string): string {
+  return str.split(/[\s_-]/)
+    .map((w) => `${w.charAt(0).toUpperCase()}${w.slice(1)}`)
+    .join(' ');
+}
