+++ src/mlmd/Utils.tsx
@@ -24,32 +24,6 @@ import {LineageResource} from "./LineageTypes";
 import {ArtifactTypeMap} from "./LineageApi";
 import {Artifact, Execution, Value} from '..';
 
-export const logger = {
-  error: (...args: any[]) => {
-    // tslint:disable-next-line:no-console
-    console.error(...args);
-  },
-  verbose: (...args: any[]) => {
-    // tslint:disable-next-line:no-console
-    console.log(...args);
-  },
-};
-
-export function formatDateString(date: Date | string | undefined): string {
-  if (typeof date === 'string') {
-    return new Date(date).toLocaleString();
-  } else {
-    return date ? date.toLocaleString() : '-';
-  }
-}
-
-/** Title cases a string by capitalizing the first letter of each word. */
-export function titleCase(str: string): string {
-  return str.split(/[\s_-]/)
-    .map((w) => `${w.charAt(0).toUpperCase()}${w.slice(1)}`)
-    .join(' ');
-}
-
 export function getResourceProperty(resource: Artifact | Execution,
     propertyName: string, fromCustomProperties = false): string | number | null {
   const props = fromCustomProperties
@@ -65,10 +39,7 @@ function getArtifactName(artifact: Artifact): string {
 }
 
 function getExecutionName(execution: Execution): string {
-  let name;
-  name = getResourceProperty(execution, ExecutionProperties.COMPONENT_ID)
-    || getResourceProperty(execution, ExecutionProperties.NAME);
-  return String(name) || '';
+  return String(getResourceProperty(execution, ExecutionProperties.NAME))
 }
 
 export function getResourceName(resource: LineageResource): string {
